refactor(reference-data): split saga postgres_registry.go by operation type
Split the 907-line registry into focused files:
- postgres_registry_read.go: GetByID, GetDefinition, GetActive, ListByStatus
- postgres_registry_write.go: CreateDraft, UpdateDefinition, ActivateSaga, DeprecateSaga
- postgres_registry_platform.go: GetPlatformSagaByID/ByName/AtTime, ComputeScriptHash, VerifyScriptHash
- postgres_registry.go: struct, constructor, transaction helpers, scan helpers
